- error dialog bug fix
