+++ app/src/main/java/am/simple/chat/core/view/BaseActivity.kt
@@ -1,10 +1,14 @@
 package am.simple.chat.core.view
 
 import am.simple.chat.R
+import am.simple.chat.core.data.type.ErrorType
 import am.simple.chat.core.utils.hasInternetConnection
+import android.content.DialogInterface
 import android.os.Bundle
 import androidx.annotation.LayoutRes
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.Observer
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.findNavController
 import androidx.navigation.fragment.FragmentNavigator
@@ -18,6 +22,9 @@ import kotlinx.coroutines.launch
 abstract class BaseActivity<ACTIVITY_VIEW_MODEL : BaseActivityViewModel> : AppCompatActivity() {
 
     protected lateinit var viewModel: ACTIVITY_VIEW_MODEL
+
+    private lateinit var dialogBuilder: AlertDialog.Builder
+    private lateinit var alertDialog: AlertDialog
 
     override fun onCreate(savedInstanceState: Bundle?) {
         initActivityViewModel()
@@ -46,7 +53,28 @@ abstract class BaseActivity<ACTIVITY_VIEW_MODEL : BaseActivityViewModel> : AppCo
 
     abstract fun setupView()
 
-    open fun initObservers() = Unit
+    open fun initObservers() {
+        viewModel.errorLiveData.observe(this, Observer {
+            showErrorDialog(errorMsgInfo = when(it.code){
+                ErrorType.ERROR_TYPE_NO_INTERNET_CONNECTION -> getString(R.string.no_internet_connection)
+                else -> it.message
+            })
+        })
+    }
+
+    private fun showErrorDialog(
+        errorMsgInfo: String? = resources.getString(R.string.something_went_wrong),
+        posBtnText: String = resources.getString(R.string.ok)
+    ) {
+        dialogBuilder = AlertDialog.Builder(this)
+            .setMessage(errorMsgInfo ?: resources.getString(R.string.something_went_wrong))
+            .setCancelable(false)
+            .setPositiveButton(posBtnText) { _: DialogInterface?, _: Int ->
+                alertDialog.dismiss()
+            }
+        alertDialog = dialogBuilder.create()
+        dialogBuilder.show()
+    }
 
     open fun loadIntentExtras() = Unit
